[UNDERTOW-2070] At ServletOutputStreamImpl, allow redirects with content length set by setting a zero content length on close even if CONTENT_LENGTH header is set. Also, still keep the "CONTENT_LENGTH is not set" restriction to prevent overwriting the value if buffer is not null.
As demonstrated by ResponseWriterTestCase.testContentLengthBasedFlush, the latter part of keeping the restriction is important to have flush taking into account a content length that is set for normal requests without redirects, where data is printed to the ServletOutputStreamImpl. This will cause the flush to limit the length of printed bytes to be equal to the content length set by the servlet.

As for the UNDERTOW-2070 bug itself, as demonstrated by RedirectWithContentLengthTestCase, a response is expected even if content length is set by the servlet before redirecting. By having ServletOutputStreamImpl.close setting the zero length in this scenario, we prevent the server from mistaking the redirect with an unfinished response, which is the cause of the empty reply.